+++ relay-server/src/processing/transactions/types.rs
@@ -1,0 +1,17 @@
+mod expanded;
+pub use expanded::ExpandedTransaction;
+mod serialized;
+pub use serialized::SerializedTransaction;
+mod output;
+pub use output::TransactionOutput;
+
+/// Flags extracted from transaction item headers.
+///
+/// Ideally `metrics_extracted` and `spans_extracted` will not be needed in the future. Unsure
+/// about `fully_normalized`.
+#[derive(Debug, Default)]
+pub struct Flags {
+    pub metrics_extracted: bool,
+    pub spans_extracted: bool,
+    pub fully_normalized: bool,
+}