+++ relay-server/src/services/processor/report.rs
@@ -269,7 +269,6 @@ fn outcome_from_parts(field: ClientReportField, reason: &str) -> Result<Outcome,
 
 #[cfg(test)]
 mod tests {
-    use relay_cogs::Token;
     use relay_config::Config;
     use relay_event_schema::protocol::EventId;
 
@@ -334,10 +333,7 @@ mod tests {
             },
         };
 
-        let envelope = processor
-            .process(&mut Token::noop(), message)
-            .await
-            .unwrap();
+        let envelope = processor.process(message).await.unwrap();
         assert!(envelope.is_none());
     }
 
@@ -393,9 +389,7 @@ mod tests {
             },
         };
 
-        let Ok(Some(Submit::Envelope(new_envelope))) =
-            processor.process(&mut Token::noop(), message).await
-        else {
+        let Ok(Some(Submit::Envelope(new_envelope))) = processor.process(message).await else {
             panic!();
         };
         let item = new_envelope.envelope().items().next().unwrap();
@@ -460,10 +454,7 @@ mod tests {
             },
         };
 
-        let envelope = processor
-            .process(&mut Token::noop(), message)
-            .await
-            .unwrap();
+        let envelope = processor.process(message).await.unwrap();
         assert!(envelope.is_none());
     }
 
@@ -502,9 +493,7 @@ mod tests {
             ctx: processing::Context::for_test(),
         };
 
-        let Ok(Some(Submit::Envelope(new_envelope))) =
-            processor.process(&mut Token::noop(), message).await
-        else {
+        let Ok(Some(Submit::Envelope(new_envelope))) = processor.process(message).await else {
             panic!();
         };
         let new_envelope = new_envelope.envelope();
@@ -552,9 +541,7 @@ mod tests {
             ctx: processing::Context::for_test(),
         };
 
-        let Ok(Some(Submit::Envelope(new_envelope))) =
-            processor.process(&mut Token::noop(), message).await
-        else {
+        let Ok(Some(Submit::Envelope(new_envelope))) = processor.process(message).await else {
             panic!();
         };
         let new_envelope = new_envelope.envelope();