+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 **Internal**:
 
 - Derive the rate limiting decision in Relay from consumed quota. ([#5390](https://github.com/getsentry/relay/pull/5390))
+- Use new processor architecture to process transactions. ([#5379](https://github.com/getsentry/relay/pull/5379))
 
 ## 25.11.0